requirements.txt
pandas==1.3.0
numpy==1.21.0
matplotlib==3.4.2
folium==0.12.1
scikit-learn==0.24.2
seaborn==0.11.1
streamlit==0.88.0
Pillow==8.2.0
